+++ tests/test_dct.c
@@ -37,10 +37,10 @@ void test_wikipedia(void) {
     };
 
     int16_t vec_freq[64];
-    double vec_temp[64];
+    float vec_temp[64];
 
     for (uint8_t i = 0; i < 64; i++) {
-        vec_temp[i] = (double)ref[i] -  128.0;
+        vec_temp[i] = (float)ref[i] -  128.0;
     }
 
     applique_dct(vec_temp);
